New add-on applicationDictionary
## Usage
This simple add-on makes possible to create application-specific dictionaries for NVDA.
To use the add-on, just go to the application for wich you want a dictionary for, and press NVDA+shift+p.
The normal NVDA dictionary dialog will open for you to add entries valid only for the active application (the application name should appear in the dialog title).
The default add-on shortcut can be changed through NVDA's input gestures dialog: under configuration node, expand the node that says "Shows the application-specific dictionary dialog".

## Command
NVDA+Shift+p

## Automatic update
This add-on includes an automatic update feature.
The check for a new version will be executed everytime NVDA is loaded.
If you do not want this, go to NVDA, Preferences, Options and in the add-on category uncheck the check box.

[1]: https://github.com/ruifontes/applicationDictionary-/releases/download/2022.03/applicationDictionary-2022.03.nvda-addon